Implement fence messages
This can be used to synchronise format changes with framebuffer updates.

This can also be used to measure network characteristics, which is
useful for implementing supplementary congestion control on the server.

A synchronisation mechanism like this isn't particularly useful with
single-threaded dispatch, but this one can be put to good use as a
measurement tool.