enc: h264: ffmpeg: Use sanctioned method for getting hw_frames_ctx
When dealing with the FFmpeg project, it's dangerous to just read the
code. You also have to read the comments because sometimes they warn you
that you're not supposed to use a particular part of the API.

Maybe that stuff should have been opaque in the first place? Oh, well...