enc: tight: Fix the size of the tile
If refinery is disabled, pixels at the tile its boundaries are not
considered by the tight encoder. This can trigger the
assert(self->n_rects > 0); when only pixels at the boundary are
damaged, or can cause those lines not to be updated. Remove the -1
from y2, x2 to fix the size of the tile.