Send queued message when switching to continues updates
Whenever a new client connects, LED status message is already
queued for the client and expected to be send in response to a
FramebufferUpdateRequest. However if the client sent one request
for a frame buffer and thereafter switches to continues updates,
a frame buffer update itself will not be send. So make sure all
pending messages are send when switching to continues updates.