stream: ws: change the Sec-WebSocket-Protocol to binary
Old noVNC versions supported both binary and base64 subprotocols and
allowed the server to pick one. The chat and the typo char are example
protocols. This will pick the binary protocol.

Since noVnc dropped this years ago it can also be argued that support
for Sec-WebSocket-Protocol can be dropped completely.